Enhanced build scripts

+++ Build/install.sh
@@ -38,29 +38,38 @@ function install_dependencies {
 }
 
 function install_typo3 {
+    print_header "Get TYPO3 source $TYPO";
     cd ..;
-    print_header "Get TYPO3 source";
-    git clone --single-branch --branch ${TYPO3} --depth 1 git://git.typo3.org/Packages/TYPO3.CMS.git;
-    cd TYPO3.CMS;
-    composer install --no-dev;
+
+    if [[ ! -e "TYPO3.CMS" ]]; then
+        git clone --single-branch --branch ${TYPO3} --depth 1 git://git.typo3.org/Packages/TYPO3.CMS.git;
+        cd TYPO3.CMS;
+    else
+        cd TYPO3.CMS;
+        git pull;
+    fi
+
+    composer install;
     rm -rf typo3/sysext/compatibility6;
 
     mkdir -p ./typo3conf/ext/;
-    ln -s ${CLI_HOME} "./typo3conf/ext/$REPO";
+    if [[ ! -e "./typo3conf/ext/$REPO" ]]; then
+        ln -s ${CLI_HOME} "./typo3conf/ext/$REPO";
+    fi
     cd ..;
 }
 
 function prepare_database {
-    $(get_mysql_client_path) -e "create database $typo3DatabaseName;" || {
-        print_info "Database $typo3DatabaseName not created";
-    };
+    if [[ "$(get_mysql_client_path)" != "" ]]; then
+        $(get_mysql_client_path) -e "create database $typo3DatabaseName;" || {
+            print_info "Database $typo3DatabaseName not created";
+        };
+    fi
 }
 
 function main {
-    echo $CLI_HOME
     cd ${CLI_HOME};
 
-    ls -hal;
     install_dependencies;
     install_typo3;
     prepare_database;
